gcc 3.4 fix

+++ trunk/avidemux/avidemux/ADM_nuv/RTjpegN.cpp
@@ -136,7 +136,7 @@ int RTjpeg::b2s(int16_t *data, int8_t *strm, uint8_t bt8)
 // return 2;
 
  // first byte allways written
- (uint8_t)strm[0]=
+ strm[0]=
       (uint8_t)(data[RTjpeg_ZZ[0]]>254) ? 254:((data[RTjpeg_ZZ[0]]<0)?0:data[RTjpeg_ZZ[0]]);
 
 
@@ -146,7 +146,7 @@ int RTjpeg::b2s(int16_t *data, int8_t *strm, uint8_t bt8)
  bitten = ((unsigned char)ci) << 2;
 
  if (ci==0) {
-   (uint8_t)strm[1]= bitten; 
+  	strm[1]= bitten; 
    co = 2;
    return (int)co;
  }
@@ -175,7 +175,7 @@ int RTjpeg::b2s(int16_t *data, int8_t *strm, uint8_t bt8)
    }
 
    if( bitoff == 0 ) {
-      (uint8_t)strm[co]= bitten; 
+     strm[co]= bitten; 
       bitten = 0;
       bitoff = 8; 
       co++;
@@ -187,7 +187,7 @@ int RTjpeg::b2s(int16_t *data, int8_t *strm, uint8_t bt8)
  /* ci must be 0 */
  if(bitoff != 6) {
 
-      (uint8_t)strm[co]= bitten; 
+      	strm[co]= bitten; 
       co++;
      
  }
@@ -204,7 +204,7 @@ HERZWEH:
    break;
  case 2:
  case 0:
-   (uint8_t)strm[co]= bitten; 
+   strm[co]= bitten; 
    bitoff = 4;
    co++;
    bitten = 0; // clear half nibble values in bitten
@@ -225,7 +225,7 @@ HERZWEH:
    bitten |= (ZZvalue&0xf)<<bitoff;
 
    if( bitoff == 0 ) {
-      (uint8_t)strm[co]= bitten; 
+      strm[co]= bitten; 
       bitten = 0;
       bitoff = 8;
       co++;
@@ -235,14 +235,14 @@ HERZWEH:
  
  /* ci must be 0 */
  if( bitoff == 0 ) {
-    (uint8_t)strm[co]= bitten; 
+    strm[co]= bitten; 
     co++;
  }  
  goto BAUCHWEH;
 
 HIRNWEH:
 
- (uint8_t)strm[co]= bitten; 
+ strm[co]= bitten; 
  co++;
